chore: switch to yarn v3 (#1752)
NPM got somehow [broken more than
usual](https://github.com/apify/crawlee/actions/runs/3997285915/jobs/6873858419)
and its no longer possible to install the dependencies in CI. We were
struggling with NPM for quite some time (especially after we moved to
turborepo). I’ve been advocating for yarn since my first days here, been
using it for last ~5 years on a daily basis and it never did anything
wrong to me, so I hope this will be the right step and a smooth
transition for all the collaborators.

What to expect? Since we require node 16, we can leverage `corepack`, a
new tool shipped with newer node versions (btw developed by yarn
maintainers). This means you just need to run `corepack enable` and yarn
should be available in your workspace. In case of some problems, you can
try running `corepack prepare yarn@stable --activate`, but it shouldn’t
be needed. Note that we use `corepack` instead of commiting yarn binary
to the git repo - that is the reason why `.yarn` folder is gitignored.

Initially we will start with a conservative setup, using a node linker,
so there will be the good `node_modules` folder we are all used to. I’d
like to experiment with the advanced features as well, mainly in the
project templates, as it could greatly improve the build times (we could
even skip the install step completely).

With that said, we will be still using NPM inside our E2E tests, to be
sure the project can be safely installed with NPM. This change should
affect only the collaborators, not our users. Ideally we should test
other package managers via CI too, so we are sure `pnpm` works as well,
or maybe even `bun`.

Few API differences and quality of life improvements yarn brings:
* `npm install` -> `yarn`
* `npm install xxx` -> `yarn add xxx`
* `npm run ...` -> `yarn …`
* `npx …` -> `yarn …`
* no need to deal with the `--` issue from `npm run` (propagating CLI
params to the script)
* powerful caching and _sane_ lockfile format