feat(MemoryStorage): read from fs if persistStorage is enabled, ram only otherwise (#1761)
This *should* just work and have no issues. Effectively
- if `persistStorage` is enabled, we don't store the data in memory as
well as on disk anymore, but rather fetch it on demand from disk
- otherwise, it is kept in memory

This does come with a caveat: when a store of any kind gets loaded and
interpreted from disk, even if persistStorage is set to false, the
existing entries will act as file system entries. This means a KVS with
an INPUT.json will still load from fs when requested and will not be
available in RAM.