feat: exceeding maxSessionRotations calls failedRequestHandler (#2029)
Any request exceeding the `maxSessionRotations` limit currently kills
the crawler. This was intended for early exit on too many hard proxy
errors, but proved to be somewhat confusing for users using
`retryOnBlocked` (any page that the crawler cannot access due to bot
protection kills the run).

With this PR, the requests that cross the limit of `maxSessionRotations`
now get processed with `failedRequestHandler`. Not sure if this is
breaking (definitely might confuse a dev or two), but it's in line with
how the crawlers worked before the `SessionError` update (so... it's
actually reverting the hidden breaking change?)

Closes #2028